+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Adds ``twine`` to dev-packages
 
+### Changed
+- Adds REAL_REPO constant
+
 ### Fixed
 - Removes useless environment variables